拆分_execute_request
